+++ rct229/rulesets/ashrae9012022/__init__.py
@@ -9,8 +9,8 @@ __all__ = ["section6"]
 
 rules_dict = {
     "PRM9012022Rule86r63": "section5rule43",
-    "PRM9012022Rule23o29": "section6rule10",
-    "PRM9012022Rule12d80": "section6rule11",
+    "PRM9012022Rule23o29": "section6rule11",
+    "PRM9012022Rule12d80": "section6rule12",
 }
 
 section_list = [
